Moved PrincipalConstants up one level.
